files (9)/requirements.txt
# Self-Healing RAG — Dependencies
# Python 3.10+

# ── Core RAG stack ────────────────────────────
chromadb==0.5.3
sentence-transformers==3.0.1
langchain==0.2.14
langchain-community==0.2.14
rank-bm25==0.2.2              # BM25 sparse retrieval

# ── PDF support ───────────────────────────────
pypdf==4.3.1

# ── LLM backends ─────────────────────────────
google-generativeai==0.7.2    # Gemini (primary)
# transformers>=4.40.0         # Uncomment for local HF fallback
# torch>=2.0.0                 # Uncomment for local inference

# ── Evaluation ────────────────────────────────
ragas==0.1.18
datasets==2.20.0

# ── API server ────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Web UI ────────────────────────────────────
streamlit==1.37.0
plotly==5.22.0

# ── Utilities ─────────────────────────────────
python-dotenv==1.0.1
tqdm==4.66.4
pydantic==2.7.4

# ── LLM provider abstraction ──────────────────────────────────────────────────
litellm>=1.40.0               # 100+ LLM providers via single interface

# ── Fine-tuning (optional — uncomment for LoRA training) ────────────────────
# trl>=0.9.0
# peft>=0.11.0
# bitsandbytes>=0.43.0
# accelerate>=0.30.0

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest-asyncio==0.23.7        # async test support
ruff==0.4.8                   # linting
black==24.4.2                 # formatting
